project_v2/iot_platform/simulator/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0